pyproject.toml
[project]
name = "ai-rag-assistant-v3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.0",
    "dagshub>=0.6.5",
    "dvc[s3]>=3.66.1",
    "fastapi>=0.128.6",
    "gradio>=6.5.1",
    "langchain>=1.2.9",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.0",
    "langchain-ollama>=1.0.1",
    "langsmith>=0.7.1",
    "mlflow>=3.9.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pymupdf>=1.26.7",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.2.2",
    "uvicorn>=0.40.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
